Update elastic_constants_dynamic.py
Reading in 'c_virial[*]' instead of 'c_stress[*]' into the stress tensor to account for changes in born_matrix.template. Added in negative sign to account for the fact that the pressure tensor from LAMMPS is the negative of the stress tensor. Added in an outer parenthesis for the fluctuation term.